bw: remove checked casts on user ids

+++ BinWeevils.GameServer/WeevilData.cs
@@ -17,9 +17,9 @@ namespace BinWeevils.GameServer
         public TypedVar<ulong> m_weevilDef;
         public TypedVar<string> m_apparel;
         public TypedVar<int> m_locID;
-        public TypedVar<int> m_doorID;
+        public TypedVar<byte> m_doorID;
         public TypedVar<int> m_poseID;
-        public TypedVar<int> m_expressionID;
+        public TypedVar<byte> m_expressionID;
         public TypedVar<int> m_victor;
         public TypedVar<double> m_x;
         public TypedVar<double> m_y;
@@ -43,9 +43,9 @@ namespace BinWeevils.GameServer
             m_weevilDef = new TypedVar<ulong>(this, "weevilDef", Var.TYPE_STRING);
             m_apparel = new TypedVar<string>(this, "apparel", Var.TYPE_STRING);
             m_locID = new TypedVar<int>(this, "locID", Var.TYPE_NUMBER);
-            m_doorID = new TypedVar<int>(this, "doorID", Var.TYPE_NUMBER);
+            m_doorID = new TypedVar<byte>(this, "doorID", Var.TYPE_NUMBER);
             m_poseID = new TypedVar<int>(this, "ps", Var.TYPE_NUMBER);
-            m_expressionID = new TypedVar<int>(this, "ex", Var.TYPE_NUMBER);
+            m_expressionID = new TypedVar<byte>(this, "ex", Var.TYPE_NUMBER);
             m_victor = new TypedVar<int>(this, "vict", Var.TYPE_STRING); // ! intentionally string
             m_x = new TypedVar<double>(this, "x", Var.TYPE_NUMBER);
             m_y = new TypedVar<double>(this, "y", Var.TYPE_NUMBER);
