+++ BinWeevils.Protocol/Str/Expression.cs
@@ -4,12 +4,12 @@ namespace BinWeevils.Protocol.Str
 {
     public partial struct ClientExpression
     {
-        [StrField] public int m_expressionID;
+        [StrField] public byte m_expressionID;
     }
     
     public partial struct ServerExpression
     {
-        [StrField] public int m_uid;
-        [StrField] public int m_expressionID;
+        [StrField] public ulong m_uid;
+        [StrField] public byte m_expressionID;
     }
 }
