+++ BinWeevils.Protocol/Str/JoinRoomRequest.cs
@@ -9,7 +9,7 @@ namespace BinWeevils.Protocol.Str
         [StrField] public double m_entryY;
         [StrField] public double m_entryZ;
         [StrField] public int m_entryDir;
-        [StrField] public int m_entryDoorID;
+        [StrField] public byte m_entryDoorID;
         [StrField] public int m_locID;
     }
 }
