+++ BinWeevils.Protocol/Str/Move.cs
@@ -11,7 +11,7 @@ namespace BinWeevils.Protocol.Str
     
     public partial struct ServerMove
     {
-        [StrField] public int m_uid;
+        [StrField] public ulong m_uid;
         [StrField] public double m_x;
         [StrField] public double m_z;
         [StrField] public int m_dir;
